requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
boto3==1.35.60
botocore==1.35.60
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.1
jmespath==1.0.1
mysqlclient==2.2.6
openai==1.54.4
pdf2image==1.17.0
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
